+++ mcp/Dockerfile
@@ -15,6 +15,7 @@
 ###
 
 FROM docker.io/library/python:3.10-slim
+
 WORKDIR /app
 
 # Install system dependencies (curl for healthcheck)
