+++ mcp/requirements.txt
@@ -14,7 +14,8 @@
 # limitations under the License.
 ###
 
-mcp[cli]>=1.6.0,<2.0.0requests>=2.26.0
+mcp[cli]>=1.6.0,<2.0.0
+requests>=2.26.0
 python-dotenv
 uvicorn>=0.34.1
 tomli>=2.0.0
